Actually wait for the online monitoring node(s) to finish setting up before continuing run setup.

+++ mnvruncontrol/backend/DataAcquisitionManager.py
@@ -1587,6 +1587,7 @@ class DataAcquisitionManager(Dispatcher.Dispatcher):
 
 		responses = None
 
+		ok_node = False
 		try:
 			responses = self.NodeSendWithResponse(message, node_type=RemoteNode.MONITORING, timeout=10)
 		except DAQErrors.NodeError:
@@ -1599,12 +1600,14 @@ class DataAcquisitionManager(Dispatcher.Dispatcher):
 				elif isinstance(response.success, Exception):
 					self.NewAlert(notice="Error starting the '%s' online monitoring node.  Error message text:\n%s\nThe run will be allowed to continue, but you will not have online monitoring..." % (response.sender, response.success), severity=Alert.ERROR)
 				else:
-					self.remote_nodes[response.sender].status = RemoteNode.OK
-					self.postoffice.Send( self.StatusReport(items=["remote_nodes"]) )
-
-					self.logger.info("    ... '%s' node started.", response.sender)
+					self.logger.info("    ... '%s' node starting.  Awaiting confirmation of full startup.", response.sender)
+					ok_node = True
 		
-		os.kill(os.getpid(), signal.SIGUSR1)
+		# if there weren't any nodes that were ok,
+		# then we don't need to wait.  just go on.
+		if not ok_node:
+			os.kill(os.getpid(), signal.SIGUSR1)
+		
 
 	def StartRemoteServices(self):
 		""" Notify all the remote services that we're ready to go.
@@ -1845,12 +1848,30 @@ class DataAcquisitionManager(Dispatcher.Dispatcher):
 			self.StopDataAcquisition()
 
 			self.postoffice.Send( self.StatusReport(items=["remote_nodes"]) )
+		elif message.state == "om_ready":
+			self.remote_nodes[message.sender].status = RemoteNode.OK
+			self.postoffice.Send( self.StatusReport(items=["remote_nodes"]) )
 
-
-		# DON'T HANDLE THE "om_ready" CASE HERE.
-		# it will deadlock because the StartOM()
-		# method is ALREADY waiting for that response!
+			self.logger.info("    ... '%s' node started.", message.sender)
 		
+			all_ready = True
+			for node_name in self.remote_nodes:
+				all_ready = all_ready and self.remote_nodes[node_name].status == RemoteNode.OK
+			
+			if all_ready:
+				self.logger.info("All monitoring nodes are ready.  Continuing in startup sequence.")
+				
+				# the main thread might still be waiting for
+				# confirmation messages that the nodes have begun
+				# the initialization (see StartOM()) -- the message
+				# receipt order is not guaranteed if the OM node
+				# starts up quickly.  so, to make sure we don't
+				# deadlock, we cancel any waiting-for-confirmation here.
+				self.postoffice.AbortMessageWait()
+				
+				# signal this process that we can go on to the next stage
+				os.kill(os.getpid(), signal.SIGUSR1)
+				
 	def PrepareRunSeries(self):
 		""" Prepares a run series based on the values
 		    found in the current configuration.  """
